+++ docs/changelog.rst
@@ -1,5 +1,16 @@
 Changelog
 =========
 
-v0.0.0 - Initial release
-------------------------
+22/03/2018
+----------
+- Deep cleaning of the package
+- Package has been moved to argh for the input parsing. All features are callable from the cli.
+- 3 options: create-files, generate-SNPs-file, regression
+- gamma regression is now callable using a `--gamma` flag in regression
+
+v1.1.0, 02/04/2019
+------------------
+
+- Genome Wide Heritability analysis available
+- Files formatted according to pep8 standards
+- Changed output strategy, now the code can be included in a snakemake pipeline